+++ packages/graphql/index.ts
@@ -1,3 +1,5 @@
+/// <reference path="./apolloTypeOverrides.d.ts" />
+
 export * from './apollo'
 export * from './components/GraphQLProvider'
 export { default as fragments } from './generated/fragments.json'