+++ packages/algolia-personalization/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @graphcommerce/algolia-personalization
 
+## 10.1.0-canary.23
+
 ## 10.1.0-canary.22
 
 ## 10.1.0-canary.21